Changed System time fields, variables, and parameters to float instead of double to work with MonoGame.

+++ Core/Objects/IUpdate.cs
@@ -4,7 +4,7 @@ namespace Atlas.Core.Objects
 {
 	public interface IUpdate : IMessenger
 	{
-		void Update(double deltaTime);
+		void Update(float deltaTime);
 		TimeStep UpdateState { get; }
 	}
 }
